requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.3.2
python-multipart==0.0.6
typing-extensions==4.9.0
starlette==0.36.3
httpx==0.26.0 